requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography
cryptography==44.0.0
distlib==0.3.8
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
filelock==3.16.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
pika==1.3.2
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
python-dotenv==1.0.1
python-jose~=3.3.0
pika==1.3.2
python-multipart==0.0.12
pytz~=2024.2
PyYAML==6.0.2
pyzmq==26.2.0
rich==13.9.1
setuptools==68.2.2
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.38.6
typer==0.12.5
typing_extensions==4.12.2
uvicorn==0.31.0
uvloop==0.20.0
virtualenv==20.26.6
watchfiles==0.24.0
websockets==13.1
pydantic-extra-types==2.10.3
ntplib==0.4.0
